First draft of using DRCellSlideActions to control tasks
This commit adds the ability to mark task work sessions in the
today's work screen as complete or incomplete by sliding the
cell to the right, as in Mail. It also changes the delete to
use the cell slide actions. I made a class to handle all of the
cell slide action stuff, which I'll use to add these actions
to the all tasks screen soon.

Note that all of the colors and such need work, and the actions
need icons, as well.