Marking a work session as complete/incomplete updates parent task
Added some logic when marking a work session as complete or
incomplete to update its parent task. If there's still work left
to do, it'll mark the parent task as incomplete, otherwise
it'll mark the parent task as complete. Might need to change this
to accommodate the case where a task with work left to be done
is marked as complete in the all tasks screen.